+++ XRViewer/ViewController/ViewController+ARSCNViewDelegate.swift
@@ -1,0 +1,17 @@
+import ARKit
+
+extension ViewController: ARSCNViewDelegate {
+    
+    func session(_ session: ARSession, cameraDidChangeTrackingState camera: ARCamera) {
+        
+        textManager.showTrackingQualityInfo(for: camera.trackingState, autoHide: true)
+        switch camera.trackingState {
+        case .notAvailable:
+            fallthrough
+        case .limited:
+            textManager.escalateFeedback(for: camera.trackingState, inSeconds: 3.0)
+        case .normal:
+            textManager.cancelScheduledMessage(forType: .trackingStateEscalation)
+        }
+    }
+}