+++ Changelog.md
@@ -1,5 +1,8 @@
 # WebXR Viewer Changelog
 
+* Added TextManager class
+* Started coordinating ARKit updates via ViewController+ARSCNViewDelegate
+* Implemented: messagePanel, messageLabel for tracking updates
 * Broke down OverlayHeader.h file into Swift constants, helper functions, and properties
 * Removed: RecordController, RecordState enum, recordDot, helperLabel, recordTimingLabel, recordButton, references to Microphone functionality
 * Dropped in Swift version of Reachability